requirements.txt
asgiref==3.9.0
Django==5.2.4
django-restframework==0.0.1
sqlparse==0.5.3
typing_extensions==4.14.0
